net: shell: Add vlan command
Move "vlan" command to a separate file.
