+++ subsys/net/lib/shell/CMakeLists.txt
@@ -29,6 +29,7 @@ zephyr_library_sources(suspend.c)
 zephyr_library_sources(tcp.c)
 zephyr_library_sources(udp.c)
 zephyr_library_sources(virtual.c)
+zephyr_library_sources(vlan.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
